clang-tidy MEDIUM: bugprone-signed-char-misuse — explicit casts + correct types
Five sites flagged for implicit signed-char widening. Each addressed at
the right level:

- menu_cbs_ok.c:4151 — `int ret = false` was assigned the int8_t return
  of config_save_overrides; the implicit narrowing path is the
  diagnostic. Declare ret as int8_t to match the function signature; the
  switch over -1/0/1 still works.
- rmsgpack.c:371 — `*s = (int8_t)tmp.u8;` is the intended msgpack
  signed-int sign-extension, but the implicit int8_t→int64_t step is
  what trips the rule. Make the widening explicit:
  `*s = (int64_t)(int8_t)tmp.u8;` and add a one-line comment so the
  intent reads as deliberate.
- xmb.c:456 — `char entry_index_offset` field stores 0/1/2 only (set
  at xmb.c:3502-3511); the `unsigned entry_idx_offset = xmb->...`
  assignment at line 1921 was the diagnostic. Change the field type to
  `unsigned char` so it lives in the correct alignment block and the
  unsigned widening is well-defined.
- menu_explore.c:303-305 — `p0 = p[0]` etc. on `const char* p`, then
  fed to TOLOWER. Cast through `(unsigned char)` to match the existing
  pattern at line 287-288 (same file, same constraint). Avoids
  sign-extension UB if `char` is signed and the byte has the high bit
  set — TOLOWER itself indexes a 256-entry table, so a sign-extended
  negative would overflow the lookup.

Build verified (Linux x86_64, no warnings introduced). Cluster spans
4 files in 4 unrelated subsystems (menu cbs, msgpack reader, xmb
driver, menu explore) — too narrow for an AUDIT-POLICY rule entry per
the 3-unrelated-files threshold (it crosses the threshold but the
diagnostic itself already exists upstream as bugprone-signed-char-misuse,
so a project-rule restating it would be noise).

+++ libretro-db/rmsgpack.c
@@ -368,7 +368,9 @@ static int rmsgpack_read_int(intfstream_t *fd, int64_t *s, size_t len)
    switch (len)
    {
       case 1:
-         *s = (int8_t)tmp.u8;
+         /* Explicit double cast: msgpack signed-int format mandates
+          * sign-extension of the 8-bit value into int64_t. */
+         *s = (int64_t)(int8_t)tmp.u8;
          break;
       case 2:
          *s = (int16_t)swap_if_little16(tmp.u16);
